requirements.txt
flask==3.1.1
flask-cors== 4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.1.0
psycopg2-binary==2.9.7
authlib==1.2.0